fix: the updates are created just before sending
