+++ triples-generator/src/main/scala/ch/datascience/triplesgenerator/eventprocessing/triplescuration/persondetails/UpdatesCreator.scala
@@ -25,23 +25,23 @@ import ch.datascience.graph.model.views.RdfResource
 import ch.datascience.rdfstore.SparqlQuery
 import ch.datascience.rdfstore.SparqlValueEncoder.sparqlEncode
 import ch.datascience.tinytypes.TinyType
-import ch.datascience.triplesgenerator.eventprocessing.triplescuration.CuratedTriples.UpdateFunction
+import ch.datascience.triplesgenerator.eventprocessing.triplescuration.CuratedTriples.CurationUpdatesGroup
 import ch.datascience.triplesgenerator.eventprocessing.triplescuration.persondetails.PersonDetailsUpdater.Person
+import eu.timepit.refined.auto._
 
 import scala.language.higherKinds
 
 private[triplescuration] class UpdatesCreator {
 
-  import eu.timepit.refined.auto._
-
   def prepareUpdates[Interpretation[_]](
-      persons:   Set[Person]
-  )(implicit ME: MonadError[Interpretation, Throwable]): List[UpdateFunction[Interpretation]] =
-    persons.toList flatMap updates
+      persons:   Person
+  )(implicit ME: MonadError[Interpretation, Throwable]): CurationUpdatesGroup[Interpretation] =
+    CurationUpdatesGroup[Interpretation](
+      name = "Persons details updates",
+      updates(persons): _*
+    )
 
-  private def updates[Interpretation[_]](
-      implicit ME: MonadError[Interpretation, Throwable]
-  ): Person => List[UpdateFunction[Interpretation]] = {
+  private def updates: Person => List[SparqlQuery] = {
     case Person(id, names, emails) =>
       List(
         namesUpdate(id, names),
@@ -50,66 +50,52 @@ private[triplescuration] class UpdatesCreator {
       ).flatten
   }
 
-  private def namesUpdate[Interpretation[_]](id: ResourceId, names: NonEmptyList[Name])(
-      implicit ME:                               MonadError[Interpretation, Throwable]
-  ) = Some {
+  private def namesUpdate(id: ResourceId, names: NonEmptyList[Name]) = Some {
     val resource = id.showAs[RdfResource]
-    UpdateFunction[Interpretation](
-      s"Updating Person $resource schema:name",
-      SparqlQuery(
-        name = "upload - person name update",
-        Set(
-          "PREFIX schema: <http://schema.org/>",
-          "PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>"
-        ),
-        s"""|DELETE { $resource schema:name ?name }
-            |${INSERT(resource, "schema:name", names.toList).getOrElse("")}
-            |WHERE { 
-            |  OPTIONAL { $resource schema:name ?maybeName }
-            |  BIND (IF(BOUND(?maybeName), ?maybeName, "nonexisting") AS ?name)
-            |}
-            |""".stripMargin
-      )
+    SparqlQuery(
+      name = "upload - person name update",
+      Set(
+        "PREFIX schema: <http://schema.org/>",
+        "PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>"
+      ),
+      s"""|DELETE { $resource schema:name ?name }
+          |${INSERT(resource, "schema:name", names.toList).getOrElse("")}
+          |WHERE { 
+          |  OPTIONAL { $resource schema:name ?maybeName }
+          |  BIND (IF(BOUND(?maybeName), ?maybeName, "nonexisting") AS ?name)
+          |}
+          |""".stripMargin
     )
   }
 
-  private def emailsUpdate[Interpretation[_]](id: ResourceId, emails: Set[Email])(
-      implicit ME:                                MonadError[Interpretation, Throwable]
-  ) = Some {
+  private def emailsUpdate(id: ResourceId, emails: Set[Email]) = Some {
     val resource = id.showAs[RdfResource]
-    UpdateFunction[Interpretation](
-      s"Updating Person $resource schema:email",
-      SparqlQuery(
-        name = "upload - person email update",
-        Set(
-          "PREFIX schema: <http://schema.org/>",
-          "PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>"
-        ),
-        s"""|DELETE { $resource schema:email ?email }
-            |${INSERT(resource, "schema:email", emails.toList).getOrElse("")}
-            |WHERE  { 
-            |  OPTIONAL { $resource schema:email ?maybeEmail }
-            |  BIND (IF(BOUND(?maybeEmail), ?maybeEmail, "nonexisting") AS ?email)
-            |}
-            |""".stripMargin
-      )
+    SparqlQuery(
+      name = "upload - person email update",
+      Set(
+        "PREFIX schema: <http://schema.org/>",
+        "PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>"
+      ),
+      s"""|DELETE { $resource schema:email ?email }
+          |${INSERT(resource, "schema:email", emails.toList).getOrElse("")}
+          |WHERE  { 
+          |  OPTIONAL { $resource schema:email ?maybeEmail }
+          |  BIND (IF(BOUND(?maybeEmail), ?maybeEmail, "nonexisting") AS ?email)
+          |}
+          |""".stripMargin
     )
   }
 
-  private def labelsDelete[Interpretation[_]](id: ResourceId)(implicit ME: MonadError[Interpretation, Throwable]) =
-    Some {
-      val resource = id.showAs[RdfResource]
-      UpdateFunction[Interpretation](
-        s"Deleting Person $resource rdfs:label",
-        SparqlQuery(
-          name = "upload - person label delete",
-          Set("PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>"),
-          s"""|DELETE { $resource rdfs:label ?label }
-              |WHERE  { $resource rdfs:label ?label }
-              |""".stripMargin
-        )
-      )
-    }
+  private def labelsDelete(id: ResourceId) = Some {
+    val resource = id.showAs[RdfResource]
+    SparqlQuery(
+      name = "upload - person label delete",
+      Set("PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>"),
+      s"""|DELETE { $resource rdfs:label ?label }
+          |WHERE  { $resource rdfs:label ?label }
+          |""".stripMargin
+    )
+  }
 
   private def INSERT[TT <: TinyType { type V = String }](resource: String,
                                                          property: String,
