+++ triples-generator/src/test/scala/ch/datascience/triplesgenerator/eventprocessing/triplescuration/forks/ForkInfoUpdaterSpec.scala
@@ -30,7 +30,6 @@ import ch.datascience.triplesgenerator.eventprocessing.CommitEvent
 import ch.datascience.triplesgenerator.eventprocessing.CommitEventProcessor.ProcessingRecoverableError
 import ch.datascience.triplesgenerator.eventprocessing.EventProcessingGenerators.{commitEvents, curationRecoverableErrors}
 import ch.datascience.triplesgenerator.eventprocessing.triplescuration.CuratedTriples
-import ch.datascience.triplesgenerator.eventprocessing.triplescuration.CuratedTriples.UpdateFunction
 import ch.datascience.triplesgenerator.eventprocessing.triplescuration.CurationGenerators._
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.matchers.should
@@ -40,24 +39,25 @@ class ForkInfoUpdaterSpec extends AnyWordSpec with MockFactory with should.Match
   import EitherT._
 
   "updateForkInfo" should {
-    "do the payload transformation and prepare update functions" in new TestCase {
+    "do the payload transformation and prepare update group" in new TestCase {
       val transformedTriples = jsonLDTriples.generateOne
       (payloadTransformer
         .transform(_: CommitEvent, _: JsonLDTriples)(_: Option[AccessToken]))
         .expects(event, givenCuratedTriples.triples, maybeAccessToken)
         .returning(rightT[IO, ProcessingRecoverableError](transformedTriples))
 
-      val updateFunctions = listOf(curationUpdateFunctions[IO]).generateOne
-      (updateFunctionsCreator
+      val updateGroup = curationUpdatesGroups[IO].generateOne
+      (updatesCreator
         .create(_: CommitEvent)(_: Option[AccessToken]))
         .expects(event, maybeAccessToken)
-        .returning(rightT[IO, ProcessingRecoverableError](updateFunctions))
+        .returning(updateGroup)
 
       forkInfoUpdater
         .updateForkInfo(event, givenCuratedTriples)
         .value
         .unsafeRunSync() shouldBe Right(
-        CuratedTriples[IO](triples = transformedTriples, updates = givenCuratedTriples.updates ++ updateFunctions)
+        CuratedTriples[IO](triples       = transformedTriples,
+                           updatesGroups = givenCuratedTriples.updatesGroups :+ updateGroup)
       )
     }
 
@@ -67,29 +67,6 @@ class ForkInfoUpdaterSpec extends AnyWordSpec with MockFactory with should.Match
         .transform(_: CommitEvent, _: JsonLDTriples)(_: Option[AccessToken]))
         .expects(event, givenCuratedTriples.triples, maybeAccessToken)
         .returning(leftT[IO, JsonLDTriples](error))
-
-      val updateFunctions = listOf(curationUpdateFunctions[IO]).generateOne
-      (updateFunctionsCreator
-        .create(_: CommitEvent)(_: Option[AccessToken]))
-        .expects(event, maybeAccessToken)
-        .returning(rightT[IO, ProcessingRecoverableError](updateFunctions))
-
-      forkInfoUpdater.updateForkInfo(event, givenCuratedTriples).value.unsafeRunSync() shouldBe Left(error)
-    }
-
-    "return a ProcessingRecoverableError if the update function creator fails" in new TestCase {
-
-      val transformedTriples = jsonLDTriples.generateOne
-      (payloadTransformer
-        .transform(_: CommitEvent, _: JsonLDTriples)(_: Option[AccessToken]))
-        .expects(event, givenCuratedTriples.triples, maybeAccessToken)
-        .returning(rightT[IO, ProcessingRecoverableError](transformedTriples))
-
-      val error = curationRecoverableErrors.generateOne
-      (updateFunctionsCreator
-        .create(_: CommitEvent)(_: Option[AccessToken]))
-        .expects(event, maybeAccessToken)
-        .returning(leftT[IO, List[UpdateFunction[IO]]](error))
 
       forkInfoUpdater.updateForkInfo(event, givenCuratedTriples).value.unsafeRunSync() shouldBe Left(error)
     }
@@ -103,44 +80,18 @@ class ForkInfoUpdaterSpec extends AnyWordSpec with MockFactory with should.Match
         .expects(event, givenCuratedTriples.triples, maybeAccessToken)
         .returning(EitherT(exception.raiseError[IO, Either[ProcessingRecoverableError, JsonLDTriples]]))
 
-      val updateFunctions = listOf(curationUpdateFunctions[IO]).generateOne
-      (updateFunctionsCreator
-        .create(_: CommitEvent)(_: Option[AccessToken]))
-        .expects(event, maybeAccessToken)
-        .returning(rightT[IO, ProcessingRecoverableError](updateFunctions))
-
       intercept[Exception] {
         forkInfoUpdater.updateForkInfo(event, givenCuratedTriples).value.unsafeRunSync()
       } shouldBe exception
     }
-
-    "fail if an error is raised when creating the updates" in new TestCase {
-
-      val transformedTriples = jsonLDTriples.generateOne
-      (payloadTransformer
-        .transform(_: CommitEvent, _: JsonLDTriples)(_: Option[AccessToken]))
-        .expects(event, givenCuratedTriples.triples, maybeAccessToken)
-        .returning(rightT[IO, ProcessingRecoverableError](transformedTriples))
-
-      val exception = exceptions.generateOne
-      (updateFunctionsCreator
-        .create(_: CommitEvent)(_: Option[AccessToken]))
-        .expects(event, maybeAccessToken)
-        .returning(EitherT(exception.raiseError[IO, Either[ProcessingRecoverableError, List[UpdateFunction[IO]]]]))
-
-      intercept[Exception] {
-        forkInfoUpdater.updateForkInfo(event, givenCuratedTriples).value.unsafeRunSync()
-      } shouldBe exception
-    }
-
   }
   private trait TestCase {
     implicit val maybeAccessToken: Option[AccessToken] = accessTokens.generateOption
     val event               = commitEvents.generateOne
     val givenCuratedTriples = curatedTriplesObjects[IO].generateOne
 
-    val payloadTransformer     = mock[PayloadTransformer[IO]]
-    val updateFunctionsCreator = mock[UpdateFunctionsCreator[IO]]
-    val forkInfoUpdater        = new ForkInfoUpdaterImpl(payloadTransformer, updateFunctionsCreator)
+    val payloadTransformer = mock[PayloadTransformer[IO]]
+    val updatesCreator     = mock[UpdatesCreator[IO]]
+    val forkInfoUpdater    = new ForkInfoUpdaterImpl(payloadTransformer, updatesCreator)
   }
 }
