+++ triples-generator/src/test/scala/ch/datascience/triplesgenerator/eventprocessing/triplescuration/persondetails/UpdatesCreatorSpec.scala
@@ -18,10 +18,8 @@
 
 package ch.datascience.triplesgenerator.eventprocessing.triplescuration.persondetails
 
-import cats.MonadError
 import cats.data.NonEmptyList
 import cats.effect.IO
-import cats.implicits._
 import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.generators.Generators._
 import ch.datascience.graph.model.GraphModelGenerators._
@@ -29,11 +27,10 @@ import ch.datascience.graph.model.users.{Email, Name, ResourceId}
 import ch.datascience.rdfstore.InMemoryRdfStore
 import ch.datascience.rdfstore.entities.Person
 import ch.datascience.rdfstore.entities.bundles.renkuBaseUrl
-import ch.datascience.triplesgenerator.eventprocessing.triplescuration.persondetails.PersonDetailsUpdater.{Person => UpdaterPerson}
+import ch.datascience.triplesgenerator.eventprocessing.triplescuration.persondetails.PersonDetailsUpdater.{Person => UpdatedPerson}
 import eu.timepit.refined.auto._
 import io.renku.jsonld.syntax._
 import org.scalatest.matchers.should
-import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 
@@ -47,13 +44,11 @@ class UpdatesCreatorSpec extends AnyWordSpec with InMemoryRdfStore with ScalaChe
 
         val personId = Person(name, email).asJsonLD.entityId.get
 
-        val updates = updatesCreator.prepareUpdates[IO](
-          Set(
-            UpdaterPerson(ResourceId(personId), NonEmptyList.of(name), Set(email))
-          )
+        val updatesGroup = updatesCreator.prepareUpdates[IO](
+          UpdatedPerson(ResourceId(personId), NonEmptyList.of(name), Set(email))
         )
 
-        updates.map(update => update().fold(throw _, runUpdate).unsafeRunSync()).sequence.unsafeRunSync()
+        updatesGroup.generateUpdates().foldF(throw _, _.runAll).unsafeRunSync()
 
         val persons = findPersons
         persons should contain theSameElementsAs Set(
@@ -69,13 +64,11 @@ class UpdatesCreatorSpec extends AnyWordSpec with InMemoryRdfStore with ScalaChe
       val name     = userNames.generateOne
       val personId = Person(name).asJsonLD.entityId.get
 
-      val updates = updatesCreator.prepareUpdates[IO](
-        Set(
-          UpdaterPerson(ResourceId(personId), NonEmptyList.of(name), emails = Set.empty)
-        )
+      val updatesGroup = updatesCreator.prepareUpdates[IO](
+        UpdatedPerson(ResourceId(personId), NonEmptyList.of(name), emails = Set.empty)
       )
 
-      updates.map(update => update().fold(throw _, runUpdate).unsafeRunSync()).sequence.unsafeRunSync()
+      updatesGroup.generateUpdates().foldF(throw _, _.runAll).unsafeRunSync()
 
       val actual = findPersons
       actual should have size 1
@@ -99,13 +92,11 @@ class UpdatesCreatorSpec extends AnyWordSpec with InMemoryRdfStore with ScalaChe
 
         val name1Updated = userNames.generateOne
 
-        val updates = updatesCreator.prepareUpdates[IO](
-          Set(
-            UpdaterPerson(ResourceId(person1Id), NonEmptyList.of(name1Updated), Set(email1))
-          )
+        val updatesGroup = updatesCreator.prepareUpdates[IO](
+          UpdatedPerson(ResourceId(person1Id), NonEmptyList.of(name1Updated), Set(email1))
         )
 
-        updates.map(update => update().fold(throw _, runUpdate).unsafeRunSync()).sequence.unsafeRunSync()
+        updatesGroup.generateUpdates().foldF(throw _, _.runAll).unsafeRunSync()
 
         findPersons should contain theSameElementsAs Set(
           (person1Id.value, Some(name1Updated.value), Some(email1.value), None),
@@ -132,13 +123,11 @@ class UpdatesCreatorSpec extends AnyWordSpec with InMemoryRdfStore with ScalaChe
 
         val name1Updated = userNames.generateOne
 
-        val updates = updatesCreator.prepareUpdates[IO](
-          Set(
-            UpdaterPerson(ResourceId(person1Id), NonEmptyList.of(name1Updated), emails = Set.empty)
-          )
+        val updatesGroup = updatesCreator.prepareUpdates[IO](
+          UpdatedPerson(ResourceId(person1Id), NonEmptyList.of(name1Updated), emails = Set.empty)
         )
 
-        updates.map(update => update().fold(throw _, runUpdate).unsafeRunSync()).sequence.unsafeRunSync()
+        updatesGroup.generateUpdates().foldF(throw _, _.runAll).unsafeRunSync()
 
         findPersons should contain theSameElementsAs Set(
           (person1Id.value, Some(name1Updated.value), None, None),
@@ -167,13 +156,11 @@ class UpdatesCreatorSpec extends AnyWordSpec with InMemoryRdfStore with ScalaChe
         val name1Updates  = nonEmptyList(userNames, minElements = 2).generateOne
         val email1Updates = nonEmptySet(userEmails, minElements = 2).generateOne
 
-        val updates = updatesCreator.prepareUpdates[IO](
-          Set(
-            UpdaterPerson(ResourceId(person1Id), name1Updates, email1Updates)
-          )
+        val updatesGroup = updatesCreator.prepareUpdates[IO](
+          UpdatedPerson(ResourceId(person1Id), name1Updates, email1Updates)
         )
 
-        updates.map(update => update().fold(throw _, runUpdate).unsafeRunSync()).sequence.unsafeRunSync()
+        updatesGroup.generateUpdates().foldF(throw _, _.runAll).unsafeRunSync()
 
         val results = findPersons
         results.filter(_._1 == person1Id.value).map(_._2) shouldBe name1Updates.map(v => Some(v.value)).toList.toSet
@@ -201,8 +188,6 @@ class UpdatesCreatorSpec extends AnyWordSpec with InMemoryRdfStore with ScalaChe
       .toSet
 
   private trait TestCase {
-    val context        = MonadError[IO, Throwable]
     val updatesCreator = new UpdatesCreator
   }
-
 }
